+++ R/imports.R
@@ -1,3 +1,2 @@
 #' @import actibase
-#' @importFrom magrittr `%>%`
 NULL