+++ NEWS.md
@@ -1,3 +1,4 @@
 # actisensorlog 0.2.0
 
 * Initial CRAN submission.
+* Removing magrittr dependency.